Detailed Data Preparation Procedures
The detailed steps used in ArcGIS Pro and ClimateWNA (cWNA) for the preliminary data processing of the MaxEnt inputs is included in this appendix. The same processing was used for both the 30m and 800m resolutions and for both the main California study area as well as the Utah prediction area. All data in GCS North American 1983 unless otherwise noted.